moved flat field functions from detector.py to a new module called flat.py, updated other modules and tests
