+++ tests/test_flat_div.py
@@ -6,6 +6,7 @@ import corgidrp
 import corgidrp.data as data
 import corgidrp.mocks as mocks
 import corgidrp.detector as detector
+import corgidrp.flat as flat
 import corgidrp.l2a_to_l2b as l2a_to_l2b
 
 old_err_tracking = corgidrp.track_individual_errors
@@ -40,7 +41,7 @@ def test_flat_div():
     assert flat_dataset.all_data[0,0,0] == 1
     
     ###### create flatfield
-    flat_frame = detector.create_flatfield(flat_dataset)
+    flat_frame = flat.create_flatfield(flat_dataset)
     # check the level of counts in flatfield is approximately correct
     assert np.mean(flat_frame.data) == pytest.approx(1, abs=1e-2)
     # check that the error is determined correctly
